feedback.c: comparison warning fixed
src/feedback.c: In function 'ndm_feedback':
src/feedback.c:48:16: warning: comparison is always false
due to limited range of data type [-Wtype-limits]
  if (env_count > SIZE_MAX / sizeof(char *)) {
                ^

+++ src/feedback.c
@@ -34,7 +34,7 @@ bool ndm_feedback(
 		return false;
 
 	unsigned int i;
-	unsigned int env_count = (ret > 0) ? 1u : 0u;
+	size_t env_count = (ret > 0) ? 1u : 0u;
 	const char **env_argv;
 
 	for (i = 0; i < ret; i++) {
